requirements.txt
allure-pytest==2.8.19
allure-python-commons==2.8.18
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
docopt==0.6.2
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
pipreqs==0.4.10
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-assume==2.3.3
pywin32==227
PyYAML==5.3.1
pyyml==0.0.2
requests==2.24.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.11
yarg==0.1.9
zipp==3.4.0
